# portless-daemon/api/openapi.yaml
openapi: 3.1.0
info:
  title: Portless local control API
  version: 10.7.0
  description: >-
    Loopback-only API shared by the Portless CLI and embedded UI. Projects are
    reusable application topologies. Environments are independently configured
    and runnable instances. Public routes use readable project, environment,
    service, and artifact names; private ownership keys are never exposed.
    Authenticated native clients may send Portless-Client-Kind as `cli` or
    `mcp`; the daemon records that fixed actor category on durable operations.
    Browser sessions cannot claim a native client kind.
servers:
  - url: http://portless.localhost/api/v1
security:
  - bearer: []
  - browserSession: []
paths:
...
  /daemon:
    get:
      operationId: daemonStatus
      description: Browser-safe identity, lifecycle state, and handoff readiness for the local daemon.
      responses:
        '200':
          description: Current daemon status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DaemonStatus' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /relay:
    get:
      operationId: relayStatus
      description: Installation and end-to-end health of clean HTTP ingress and scoped TCP endpoint DNS.
      responses:
        '200':
          description: Current privileged relay status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RelayStatus' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /projects:
    get:
      operationId: listProjects
      parameters:
        - { $ref: '#/components/parameters/Limit1000' }
      responses:
        '200':
          description: Logical projects with environment summaries
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      operationId: createProject
      description: Discover and combine one or more named source checkouts, then create the local environment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, sources]
              properties:
                name: { $ref: '#/components/schemas/ProjectName' }
                sources:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/SourceInput' }
      responses:
        '201': { description: Project, initial environment, and discovery warnings }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /projects/{projectName}/sources:
    parameters:
      - { $ref: '#/components/parameters/ProjectName' }
    post:
      operationId: addProjectSource
      description: >-
        Discovers a local source directory, adds it to the logical project, and binds its absolute
        path only to the selected environment. Every environment in the project must be stopped;
        other environments remain unbound until configured explicitly. Available to authenticated
        CLI and browser clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, path, environment]
              properties:
                name: { $ref: '#/components/schemas/SourceName' }
                path: { type: string, description: Absolute source path on the daemon host. }
                environment: { $ref: '#/components/schemas/EnvironmentName' }
      responses:
        '201': { description: Updated project, selected environment, discovery warnings, and environment selectors requiring configuration }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Mode-0600 installation secret used by the local CLI.
    browserSession:
      type: apiKey
      in: cookie
      name: portless_session
      description: SameSite=Strict browser session; mutations also require X-Portless-CSRF.
  parameters:
...
    ProjectName:
      in: path
      name: projectName
      required: true
      schema: { $ref: '#/components/schemas/ProjectName' }
...
    Limit1000:
      in: query
      name: limit
      description: Positive result bound; the endpoint applies its documented default when omitted.
      schema: { type: integer, minimum: 1, maximum: 1000 }
...
  responses:
...
    Conflict:
      description: Name, revision, provider, or current-state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unavailable:
      description: Required local runtime unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ProjectName: &dnsName
      type: string
      pattern: '^[a-z][a-z0-9-]{0,62}$'
      examples: [billing]
...
    DaemonStatus:
      type: object
      required: [state, pid, startedAt, instanceId, buildId, protocolVersion, apiVersion, handoffReady, recoveryProblems, activeEnvironments]
      properties:
        state: { type: string, enum: [reconciling, ready, draining] }
        pid: { type: integer }
        startedAt: { type: string, format: date-time }
        instanceId: { type: string }
        buildId: { type: string }
        protocolVersion: { type: string, pattern: '^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)$', example: '3.0.0', description: Stable semantic daemon protocol version. }
        apiVersion: { type: string, pattern: '^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)$', example: '10.7.0', description: Stable semantic API version. }
        handoffReady: { type: boolean }
        recoveryProblems: { type: array, items: { type: string } }
        activeEnvironments: { type: array, items: { type: string, example: billing/local } }
...
    RelayStatus:
      type: object
      required: [platform, service, installed, running, healthy, httpHealthy, helperPresent, helperCurrent, dnsHealthy, resolverPresent, resolverHealthy, endpointPoolReady, dnsListenAddress]
      properties:
        platform: { type: string, enum: [launchd, systemd, unsupported] }
        service: { type: string }
        installed: { type: boolean }
        running: { type: boolean }
        healthy: { type: boolean }
        httpHealthy: { type: boolean }
        helperPresent: { type: boolean }
        helperCurrent: { type: boolean, description: Whether the installed relay helper matches the current Portless executable. }
        helperBuildId: { type: string, description: SHA-256 identity of the installed relay helper. }
        currentBuildId: { type: string, description: SHA-256 identity of the current Portless executable. }
        dnsHealthy: { type: boolean }
        resolverPresent: { type: boolean }
        resolverHealthy: { type: boolean }
        endpointPoolReady: { type: boolean }
        endpointPoolDetail: { type: string }
        targetSocket: { type: string }
        dnsTargetSocket: { type: string }
        dnsListenAddress: { type: string, example: '127.77.0.1:1053' }
        resolverPath: { type: string }
        healthError: { type: string }
        dnsHealthError: { type: string }
        resolverHealthError: { type: string }
        problem: { type: string }
    SourceInput:
      type: object
      required: [name, path]
      properties:
        name: { $ref: '#/components/schemas/SourceName' }
        path: { type: string, description: Absolute path resolved by the CLI. }
    ProjectSource:
      type: object
      required: [name]
      properties:
        name: { $ref: '#/components/schemas/SourceName' }
        services: { type: array, items: { $ref: '#/components/schemas/ServiceName' } }
...
    HealthCheck:
      type: object
      required: [kind, timeout, interval]
      properties:
        kind: { type: string, enum: [tcp, http, exec] }
        path: { type: string }
        timeout: { type: integer, description: Nanoseconds as encoded by Go time.Duration. }
        interval: { type: integer, description: Nanoseconds as encoded by Go time.Duration. }
    ServiceDefinition:
      type: object
      required: [name, kind, required, health]
      properties:
        name: { $ref: '#/components/schemas/ServiceName' }
        kind: { type: string, enum: [process, resource] }
        framework: { type: string }
        resource:
          type: object
          required: [type, version]
          properties:
            type: { type: string, example: postgres }
            version: { type: string, example: '17' }
        command: { type: array, items: { type: string } }
        workingDirectory: { type: string }
        serviceDirectory: { type: string }
        debug:
          type: object
          required: [adapter, launcher, command]
          properties:
            adapter: { type: string, enum: [node-inspector, jdwp] }
            launcher: { type: string, enum: [node-direct, nest-cli, spring-gradle, spring-maven] }
            command: { type: array, items: { type: string } }
        portEnvironment: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535, description: Plugin-resolved client-facing port for a TCP service. }
        environment: { type: object, additionalProperties: { type: string } }
        required: { type: boolean }
        health: { $ref: '#/components/schemas/HealthCheck' }
...
    Connection:
      type: object
      required: [source, target, protocol, required]
      properties:
        source: { type: string, description: Service name or external. }
        target: { $ref: '#/components/schemas/ServiceName' }
        protocol: { type: string, enum: [http, tcp] }
        binding: { type: string, description: Resource plugin responsible for connection formatting. }
        environment: { type: string }
        required: { type: boolean }
...
    EnvironmentSummary:
      type: object
      required: [project, name, revision, status, serviceCount, readyCount, remoteCount, updatedAt]
      properties:
        project: { $ref: '#/components/schemas/ProjectName' }
        name: { $ref: '#/components/schemas/EnvironmentName' }
        revision: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/EnvironmentStatus' }
        reason: { type: string }
        serviceCount: { type: integer }
        readyCount: { type: integer }
        remoteCount: { type: integer }
        updatedAt: { type: string, format: date-time }
        dashboardUrl: { type: string, format: uri }
    Project:
      type: object
      required: [name, revision, createdAt, updatedAt]
      properties:
        name: { $ref: '#/components/schemas/ProjectName' }
        revision: { type: integer, format: int64 }
        primaryService: { $ref: '#/components/schemas/ServiceName' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        dashboardUrl: { type: string, format: uri }
        sources: { type: array, items: { $ref: '#/components/schemas/ProjectSource' } }
        services: { type: array, items: { $ref: '#/components/schemas/ServiceDefinition' } }
        connections: { type: array, items: { $ref: '#/components/schemas/Connection' } }
        environments: { type: array, items: { $ref: '#/components/schemas/EnvironmentSummary' } }
...
    EnvironmentStatus:
      type: string
      enum: [starting, recovering, healthy, degraded, failed, stopping, stopped, unknown]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            subject: { type: object, additionalProperties: true }
            details: { type: object, additionalProperties: true }
            remediation: { type: array, items: { type: object, additionalProperties: true } }